Migrate formatTimeAgo over into utils/formatters and delete shared/formatters.test.js
- use toDate rather than _d since _d is not meant to be publicly accessed
- delete shared/formatters.test.js since there are no more tests in that file
