+++ src/shared/formatters.js
@@ -1,5 +1,4 @@
 import { isFinite } from 'lodash';
-import moment from 'moment';
 
 // Format a number of cents into a string, e.g. $12,345.67
 export function formatCents(cents) {
@@ -11,13 +10,4 @@ export function formatCents(cents) {
     minimumFractionDigits: 2,
     maximumFractionDigits: 2,
   });
-}
-
-export function formatTimeAgo(date) {
-  if (!date) return undefined;
-
-  return moment(date)
-    .fromNow()
-    .replace('minute', 'min')
-    .replace(/a min\s/, '1 min ');
 }
