--- src/shared/formatters.test.js
@@ -1,18 +1,0 @@
-import * as formatters from './formatters';
-import moment from 'moment';
-
-describe('formatters', () => {
-  describe('formatTimeAgo', () => {
-    it('should account for 1 minute correctly', () => {
-      let time = new Date();
-      let formattedTime = formatters.formatTimeAgo(time);
-
-      expect(formattedTime).toEqual('a few seconds ago');
-
-      time = moment().subtract(1, 'minute')._d;
-      formattedTime = formatters.formatTimeAgo(time);
-
-      expect(formattedTime).toEqual('1 min ago');
-    });
-  });
-});
