+++ src/utils/formatters.test.js
@@ -1,3 +1,5 @@
+import moment from 'moment';
+
 import * as formatters from './formatters';
 
 import PAYMENT_REQUEST_STATUS from 'constants/paymentRequestStatus';
@@ -278,6 +280,20 @@ describe('formatters', () => {
     it('should include the timezone shortcode', () => {
       const formattedDate = formatters.formatDateTimeWithTZ(new Date());
       expect(formattedDate).toMatch(/\d{2}-\w{3}-\d{2} \d{2}:\d{2} \w{2,3}/);
+    });
+  });
+
+  describe('formatTimeAgo', () => {
+    it('should account for 1 minute correctly', () => {
+      let time = new Date();
+      let formattedTime = formatters.formatTimeAgo(time);
+
+      expect(formattedTime).toEqual('a few seconds ago');
+
+      time = moment().subtract(1, 'minute').toDate();
+      formattedTime = formatters.formatTimeAgo(time);
+
+      expect(formattedTime).toEqual('1 min ago');
     });
   });
 });
